tiny_llava_requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.41.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.3
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
gradio==3.35.2
gradio_client==0.2.9
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==2.2.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
orjson==3.10.0
packaging==24.0
pandas==2.2.1
peft==0.4.0
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
pydantic==1.10.15
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12
threadpoolctl==3.4.0
tiktoken==0.6.0
timm==0.6.13
# Editable install with no version control (tinyllava==1.0.0)
-e git/mllms/TinyLLaVABench
tokenizers==0.15.1
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.29.0
wavedrom==2.0.3.post3
websockets==12.0
yarl==1.9.4
